+++ vue/components/ui/molecules/table/table.vue
@@ -139,6 +139,10 @@
     background-color: $selected-dark-color;
 }
 
+.table tbody tr.checked {
+    background-color: $selected-dark-color;
+}
+
 .table th {
     color: $label-color;
     font-size: 11px;
@@ -519,6 +523,7 @@ export const Table = {
         },
         rowClasses(item) {
             const base = {
+                checked: this.checkedItemsData[item.id],
                 selected: this.isRowSelected(item.id),
                 clickable: item.clickable === undefined ? this.clickableRows : item.clickable,
                 hoverable: item.hoverable === undefined ? this.hoverableRows : item.hoverable
@@ -530,10 +535,7 @@ export const Table = {
         },
         onClick(item, index) {
             if (this.rowSelection) {
-                this.selectedRowData =
-                    this.selectedIdData === null || this.selectedRowData !== item.id
-                        ? item.id
-                        : null;
+                this.selectedRowData = this.selectedRowData !== item.id ? item.id : null;
             }
 
             this.$emit("click", item, item._originalIndex, index, this.selectedRowData);